requirements.txt
mcp>=0.1.0
jira>=3.5.1
python-dotenv>=1.0.0
httpx>=0.24.0
fastapi>=0.104.1
uvicorn>=0.24.0 